project-3-alloydb-nl/requirements.txt
flask
sqlalchemy
pg8000
python-dotenv
google-cloud-aiplatform
vertexai
